fix: games didn't showed decimal point

+++ src/main/scala/hm/moe/pokkedoll/warscore/games/TeamDeathMatch.scala
@@ -634,7 +634,7 @@ class TeamDeathMatch(override val id: String) extends Game {
 
     comp.append("* ")
       .append("K/D: ").color(ChatColor.GRAY)
-      .append((data.kill / (data.death + 1)).toString).color(ChatColor.GREEN).bold(true)
+      .append((data.kill / (data.death + 1).toDouble).toString).color(ChatColor.GREEN).bold(true)
       .append("\n").color(ChatColor.RESET).bold(false)
     /*
         comp.append("* ")
